system: rtc: correctly deal with glib io watches and channels
This fixes a rather serios problem with rtc poll handling. When a watch is added to the
glib mainloop it references internally a file descriptor to pass it later to the poll
function (see poll(2)). Every watch gets an id which should be used by the user to refer
to the watch whenever needed. When the io channel is destroyed the watch should be remove
from the mainloop as well. If this isn't done probably by the user it will cause the
internal list of fds which are passed to the poll function to exceed the maximum of 256.
This finally results in poll returning EINVAL as error code.

2012-11-09T19:24:24.809509Z [5146] tuna user.warning sleepd: poll(2) failed due to:
Invalid argument.

GLib doesn't handle this very well and will take 100% CPU usage when this occurs.

This patch adds the correct removal of the watch from the glib mainloop when it's not
needed anymore. Furthermore the watch isn't added again when it or the channel is still
valid.

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>